AdminEdits refactor
