+++ src/redux/reducers/nextButtonReducer.js
@@ -1,7 +1,7 @@
 import { combineReducers } from "redux";
 
 // Stores NextButton and BackButton "state"
-const nextButton = (state = 0, action) => {
+const nextButton = (state = -1, action) => {
   console.log(`What is action in next button reducer`, state);
   switch (action.type) {
     case "CLEAR_NEXT_BUTTON":
